Added removing medias form different playlists. File will be removed only if deletion was from default playlists (with all files).
Little refactioring. Added constants.

+++ youtubetomp3/models.py
@@ -1,15 +1,17 @@
 from django.db import models
 from django.contrib.auth.models import User
-
 import os
 
+from youtubetomp3.const.constants import _Const
 import youtubetomp3.core.utils as Utils
+
+CONST = _Const()
 
 class PlaylistManager(models.Manager):
     """PlaylistManager with checking on name"""
     def create_playlist(self, name, user, is_audio):
         if self.filter(name=name, user=user, is_audio=is_audio).count() != 0:
-            return "DUBLICATE"
+            return CONST.DUBLICATE
         else:
             return self.create(user=user, name=name, is_audio=is_audio)
 
@@ -32,7 +34,7 @@ class MediaManager(models.Manager):
     """MediaManager with checking on name"""
     def create_media(self, playlist, mediafile):
         if self.filter(playlist=playlist, mediafile=mediafile).count() != 0:
-            return "DUBLICATE"
+            return CONST.DUBLICATE
         else:
             return self.create(playlist=playlist, mediafile=mediafile)    
         
@@ -50,11 +52,12 @@ class Media(models.Model):
         return self.playlist.user.username + ": " + self.mediafile
 
     def delete(self):
-        print self.mediafile
 
-        os.remove(Utils.createPath(
-            os.path.basename(self.mediafile), 
-            self.playlist.user))
+        if self.playlist.name == CONST.DEFAULT_AUDIO_PLAYLIST or self.playlist.name == CONST.DEFAULT_VIDEO_PLAYLIST:
+
+            os.remove(Utils.createPath(
+                os.path.basename(self.mediafile), 
+                self.playlist.user))
         
         super(Media, self).delete()
